# course-work/implementations/freelance-api/src/main/java/com/freelance/freelance_api/entities/Role.java
package com.freelance.freelance_api.entities;

import jakarta.persistence.*;
import jakarta.validation.constraints.NotBlank;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Entity
@Table(name = "roles")
@Getter
@Setter
@NoArgsConstructor
@EqualsAndHashCode(onlyExplicitlyIncluded = true)
public class Role {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @NotBlank(message = "Role name can't be empty")
    @Column(nullable = false, unique = true)
    @EqualsAndHashCode.Include
    private String name;

    public Role(String name) {
        this.name = name;
    }
}